requirements.txt
beautifulsoup4
PySide6-Fluent-Widgets[full]
asyncslot
aiohttp_retry
